# server/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: 小确幸服务端 API
  version: 0.1.0
  description: |
    S0–S3 surface: health, auth, personal space, records, sync, photo STS upload.
    Field names are lowerCamelCase. Timestamps are UTC RFC3339.
    Server IDs are UUIDv7 strings.
    Media quota: 200MB/user, 5MB/photo, 9 photos/record. Invite: link only, 7 days, max 6 members. Account delete cooling: 24h.
servers:
  - url: /
paths:
...
  /api/v1/account:
    delete:
      operationId: deleteAccount
      summary: Mark the account PENDING_DELETE and revoke sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pending delete
        "401":
          $ref: "#/components/responses/Error"
  /api/v1/me:
    get:
      operationId: getMe
      security: [{ bearerAuth: [] }]
      summary: Current account. Phone is masked.
      responses:
        "200": { description: Profile }
    patch:
      operationId: patchMe
      security: [{ bearerAuth: [] }]
      summary: Update display name (1-20 chars)
      responses:
        "200": { description: Profile }
...
  /api/v1/records/{id}:
    get:
      operationId: getRecord
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Record }
    patch:
      operationId: patchRecord
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Applied }
        "409": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteRecord
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Tombstone }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    Meta:
      type: object
      properties:
        requestId: { type: string }
        serverTime: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message, retryable]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
            retryable: { type: boolean }
        meta: { $ref: "#/components/schemas/Meta" }
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"